requirements.txt
Flask==2.3.2
Pillow==10.0.1
pytesseract==0.3.10